+++ content_manager_wrapper.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"net/url"
+	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -19,10 +20,14 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// contentManagerWrapperSeparator is a special character used by Content Manager to determine which port
-// the content manager wrapper is running on. The server name shows "<server name> <separator><port>",
-// which Content Manager uses to split the string and find the port.
-const contentManagerWrapperSeparator = 'ℹ'
+const (
+	// contentManagerWrapperSeparator is a special character used by Content Manager to determine which port
+	// the content manager wrapper is running on. The server name shows "<server name> <separator><port>",
+	// which Content Manager uses to split the string and find the port.
+	contentManagerWrapperSeparator = 'ℹ'
+
+	ContentManagerJoinLinkBase string = "https://acstuff.ru/s/q:race/online/join"
+)
 
 type ContentManagerWrapperData struct {
 	ACHTTPSessionInfo
@@ -455,4 +460,26 @@ func geoIP() (*GeoIP, error) {
 	}
 
 	return geoIPData, nil
+}
+
+func getContentManagerJoinLink(config ServerConfig) (*url.URL, error) {
+	geoIP, err := geoIP()
+
+	if err != nil {
+		return nil, err
+	}
+
+	cmUrl, err := url.Parse(ContentManagerJoinLinkBase)
+
+	if err != nil {
+		return nil, err
+	}
+
+	queryString := cmUrl.Query()
+	queryString.Set("ip", geoIP.IP)
+	queryString.Set("httpPort", strconv.Itoa(config.GlobalServerConfig.HTTPPort))
+
+	cmUrl.RawQuery = queryString.Encode()
+
+	return cmUrl, nil
 }