+++ scheduled_races.go
@@ -86,37 +86,76 @@ func BuildICalEvent(event ScheduledEvent) *components.Event {
 type ScheduledRacesHandler struct {
 	*BaseHandler
 
-	store               Store
-	raceManager         *RaceManager
-	championshipManager *ChampionshipManager
+	scheduledRacesManager *ScheduledRacesManager
 }
 
-func NewScheduledRacesHandler(baseHandler *BaseHandler, store Store, raceManager *RaceManager, championshipManager *ChampionshipManager) *ScheduledRacesHandler {
+func NewScheduledRacesHandler(baseHandler *BaseHandler, scheduledRacesManager *ScheduledRacesManager) *ScheduledRacesHandler {
 	return &ScheduledRacesHandler{
-		BaseHandler:         baseHandler,
-		store:               store,
-		raceManager:         raceManager,
-		championshipManager: championshipManager,
+		BaseHandler:           baseHandler,
+		scheduledRacesManager: scheduledRacesManager,
 	}
 }
 
 func (rs *ScheduledRacesHandler) calendar(w http.ResponseWriter, r *http.Request) {
-	rs.viewRenderer.MustLoadTemplate(w, r, "calendar.html", map[string]interface{}{
-		"WideContainer": true,
-	})
+	rs.viewRenderer.MustLoadTemplate(w, r, "calendar.html", &BaseTemplateVars{WideContainer: true})
 }
 
 func (rs *ScheduledRacesHandler) calendarJSON(w http.ResponseWriter, r *http.Request) {
 	err := rs.generateJSON(w, r)
 
 	if err != nil {
-		logrus.Errorf("could not find scheduled events, err: %s", err)
+		logrus.WithError(err).Errorf("could not find scheduled events")
 		return
 	}
 }
 
-func (rs *ScheduledRacesHandler) getScheduledRaces() ([]ScheduledEvent, error) {
-	_, _, _, customRaces, err := rs.raceManager.ListCustomRaces()
+func (rs *ScheduledRacesHandler) allScheduledRacesICalHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Add("Content-Type", "text/calendar; charset=utf-8")
+	w.Header().Add("Content-Disposition", "inline; filename=championship.ics")
+
+	err := rs.scheduledRacesManager.buildScheduledRaces(w)
+
+	if err != nil {
+		logrus.WithError(err).Error("could not build scheduled races feed")
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+}
+
+func (rs *ScheduledRacesHandler) generateJSON(w http.ResponseWriter, r *http.Request) error {
+	start, err := time.Parse(time.RFC3339, r.URL.Query().Get("start"))
+
+	if err != nil {
+		return err
+	}
+
+	end, err := time.Parse(time.RFC3339, r.URL.Query().Get("end"))
+
+	if err != nil {
+		return err
+	}
+
+	calendarObjects, err := rs.scheduledRacesManager.buildCalendar(start, end)
+
+	if err != nil {
+		return err
+	}
+
+	return json.NewEncoder(w).Encode(calendarObjects)
+}
+
+type ScheduledRacesManager struct {
+	store Store
+}
+
+func NewScheduledRacesManager(store Store) *ScheduledRacesManager {
+	return &ScheduledRacesManager{
+		store: store,
+	}
+}
+
+func (srm *ScheduledRacesManager) getScheduledRaces() ([]ScheduledEvent, error) {
+	customRaces, err := srm.store.ListCustomRaces()
 
 	if err != nil {
 		return nil, err
@@ -125,10 +164,14 @@ func (rs *ScheduledRacesHandler) getScheduledRaces() ([]ScheduledEvent, error) {
 	var scheduled []ScheduledEvent
 
 	for _, race := range customRaces {
+		if race.Scheduled.IsZero() {
+			continue
+		}
+
 		scheduled = append(scheduled, race)
 	}
 
-	championships, err := rs.championshipManager.ListChampionships()
+	championships, err := srm.store.ListChampionships()
 
 	if err != nil {
 		return nil, err
@@ -148,8 +191,8 @@ func (rs *ScheduledRacesHandler) getScheduledRaces() ([]ScheduledEvent, error) {
 	return scheduled, nil
 }
 
-func (rs *ScheduledRacesHandler) buildScheduledRaces(w io.Writer) error {
-	scheduled, err := rs.getScheduledRaces()
+func (srm *ScheduledRacesManager) buildScheduledRaces(w io.Writer) error {
+	scheduled, err := srm.getScheduledRaces()
 
 	if err != nil {
 		return err
@@ -173,20 +216,6 @@ func (rs *ScheduledRacesHandler) buildScheduledRaces(w io.Writer) error {
 
 	return err
 }
-
-func (rs *ScheduledRacesHandler) allScheduledRacesICalHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Add("Content-Type", "text/calendar; charset=utf-8")
-	w.Header().Add("Content-Disposition", "inline; filename=championship.ics")
-
-	err := rs.buildScheduledRaces(w)
-
-	if err != nil {
-		logrus.WithError(err).Error("could not build scheduled races feed")
-		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-		return
-	}
-}
-
 type CalendarObject struct {
 	ID               string    `json:"id"`
 	GroupID          string    `json:"groupId"`
@@ -211,24 +240,11 @@ type CalendarObject struct {
 	TextColor       string `json:"textColor"`
 }
 
-func (rs *ScheduledRacesHandler) generateJSON(w http.ResponseWriter, r *http.Request) error {
-
-	start, err := time.Parse(time.RFC3339, r.URL.Query().Get("start"))
-
-	if err != nil {
-		return err
-	}
-
-	end, err := time.Parse(time.RFC3339, r.URL.Query().Get("end"))
+func (srm *ScheduledRacesManager) buildCalendar(start time.Time, end time.Time) ([]CalendarObject, error) {
+	scheduled, err := srm.getScheduledRaces()
 
 	if err != nil {
-		return err
-	}
-
-	scheduled, err := rs.getScheduledRaces()
-
-	if err != nil {
-		return err
+		return nil, err
 	}
 
 	var calendarObjects []CalendarObject
@@ -269,7 +285,7 @@ func (rs *ScheduledRacesHandler) generateJSON(w http.ResponseWriter, r *http.Req
 			rule, err := customRace.GetRecurrenceRule()
 
 			if err != nil {
-				return err
+				return nil, err
 			}
 
 			for _, startTime := range rule.Between(start, end, true) {
@@ -288,10 +304,10 @@ func (rs *ScheduledRacesHandler) generateJSON(w http.ResponseWriter, r *http.Req
 
 	scheduled = append(scheduled, recurring...)
 
-	return BuildCalObject(w, scheduled, calendarObjects)
+	return BuildCalObject(scheduled, calendarObjects)
 }
 
-func BuildCalObject(w http.ResponseWriter, scheduled []ScheduledEvent, calendarObjects []CalendarObject) error {
+func BuildCalObject(scheduled []ScheduledEvent, calendarObjects []CalendarObject) ([]CalendarObject, error) {
 	for _, scheduledEvent := range scheduled {
 
 		var prevSessionTime time.Duration
@@ -404,7 +420,7 @@ func BuildCalObject(w http.ResponseWriter, scheduled []ScheduledEvent, calendarO
 		}
 	}
 
-	return json.NewEncoder(w).Encode(calendarObjects)
+	return calendarObjects, nil
 }
 
 func GenerateSummary(raceSetup CurrentRaceConfig, eventType string) string {