+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v1.x.x
+------
+
+Added:
+
+* Discord integration! Thanks to @cheesegrits for this! Check out the Server Options page for more information.
+* Dark Theme! You can now set Server Manager to use a Dark Theme in the Server Options. You can also edit this for your account in the "Update Details" page.
+
 v1.4.2
 ------